add interface for docs
