+++ public/main.js
@@ -1,0 +1,3 @@
+/**
+ * Created by rlopes on 26/12/2016.
+ */
